blinker.signal() and blinker.Namespace no longer use weak references.
In the original implementation, I was uneasy about releasing a library
that held an unbounded, module-level cache, so it was built using a
weak value mapping. In practice, users making tons of signals are
using Signal() directly, and the weak referencing violates the
principle of least surprise in code like `signal('foo').connect(...)`.

Previous code is available as WeakNamespace for the time being. The
implementation is trivial and will likely be dropped in the future
unless use cases are voiced.