Expose default namespace object.
This is useful for when you still only have a single global namespace
and want to use the global `signal` helper function, but want to be able
to modify or inspect that namespace. This also makes it possible to
write consistent code that is parameterized on a namespace, but can
still function with the global namespace. This also enables support for
diagnostic tools that inspect the global namespace.