simplify weakref handling
Use the built-in `WeakMethod` and `inspect.ismethod` instead of a custom implementation.
Build a cleanup callbacks from an id, rather than storing the id on an `annnotatable_weakref` subclass.
Simplify the private `Symbol` class implementation.
Use the built-in `cached_property` instead of a custom implementation.
Use positional-only argument for `send(sender)` instead of `*args` hack.
This simplified the type annotations, which pass with pyright as well now.